User 모델과 Post 모델을 1:N으로 연결

+++ posts/views.py
@@ -26,13 +26,21 @@ def list(request):
     
     
 def delete(request, post_id):
-    post = Post.objects.get(id=post_id)
+    post = get_object_or_404(Post, id=post_id)
+    
+    if post.user != request.user:
+        return redirect('posts:list')
+        
     post.delete()
     return redirect('posts:list')
 
 
 def update(request, post_id):
     post = get_object_or_404(Post, id=post_id)
+    
+    if post.user != request.user:
+        return redirect('posts:list')
+        
     if request.method == 'POST':
         form = PostModelForm(request.POST, instance=post)
         if form.is_valid():
